Drop use of GroupsMap
